+++ app/src/main/res/layout-land/activity_main.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:padding="10dp"
-    tools:context=".MyVisitsActivity">
+    tools:context=".MainActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
@@ -36,7 +36,7 @@
                     android:textSize="16sp" />
 
                 <ListView
-                    android:id="@+id/countries_list_visits"
+                    android:id="@+id/countries_list_main"
                     android:layout_width="200dp"
                     android:layout_height="match_parent"
                     android:layout_weight="1" />
@@ -57,7 +57,7 @@
                     android:textSize="16sp" />
 
                 <ListView
-                    android:id="@+id/provinces_list_visits"
+                    android:id="@+id/provinces_list_main"
                     android:layout_width="200dp"
                     android:layout_height="match_parent"
                     android:layout_weight="1" />
@@ -78,7 +78,7 @@
                     android:textSize="16sp" />
 
                 <ListView
-                    android:id="@+id/places_list_visits"
+                    android:id="@+id/places_list_main"
                     android:layout_width="200dp"
                     android:layout_height="match_parent"
                     android:layout_weight="1" />
@@ -87,7 +87,7 @@
         </LinearLayout>
 
         <ListView
-            android:id="@+id/visit_list"
+            android:id="@+id/visit_list_main"
             android:layout_width="match_parent"
             android:layout_height="match_parent" />
